+++ NEWS.md
@@ -2,3 +2,26 @@
 
 * classe E5R.Architecture.Core.**SemVer** básica
 * classe E5R.Architecture.Data.**Data** temporária
+* Formatador de string CPF
+```csharp
+using E5R.Architecture.Core.Formatters;
+
+// code block
+{
+   string cpfFormatado = (FormattedCPF)"12345678901"; 
+}
+
+// output $cpfFormatado => "123.456.789-01"
+```
+
+* Desformatador de string de CPF
+```csharp
+using E5R.Architecture.Core.Formatters;
+
+// code block
+{
+   string cpfDesformatado = (UnformattedCPF)"123.456.789-01"; 
+}
+
+// output $cpfDesformatado => "12345678901"
+```